+++ tests/unit/test_transforms.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import argparse
 import logging as log
+import os
 import os.path as osp
 import random
 from unittest import TestCase
@@ -14,6 +15,7 @@ from pandas.api.types import CategoricalDtype
 
 import datumaro.plugins.transforms as transforms
 import datumaro.util.mask_tools as mask_tools
+from datumaro.components.algorithms.hash_key_inference.explorer import Explorer
 from datumaro.components.annotation import (
     AnnotationType,
     Bbox,
@@ -1673,3 +1675,57 @@ class CleanTest(TestCase):
             result_item = result.__getitem__(i)
             self.assertEqual(expected_item.annotations, result_item.annotations)
             self.assertEqual(expected_item.media, result_item.media)
+
+
+class PseudoLabelingTest(TestCase):
+    def setUp(self):
+        self.data_path = get_test_asset_path("explore_dataset")
+        self.categories = ["bird", "cat", "dog", "monkey"]
+        self.source = Dataset.from_iterable(
+            [
+                DatasetItem(
+                    id=0,
+                    media=Image.from_file(
+                        path=os.path.join(self.data_path, "dog", "ILSVRC2012_val_00001698.JPEG")
+                    ),
+                ),
+                DatasetItem(
+                    id=1,
+                    media=Image.from_file(
+                        path=os.path.join(self.data_path, "cat", "ILSVRC2012_val_00004894.JPEG")
+                    ),
+                ),
+            ],
+            categories=self.categories,
+        )
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_transform_pseudolabeling_with_labels(self):
+        dataset = self.source
+        labels = self.categories
+        explorer = Explorer(dataset)
+        result = dataset.transform("pseudo_labeling", labels=labels, explorer=explorer)
+
+        label_indices = dataset.categories()[AnnotationType.label]._indices
+        for item, expected in zip(result, ["dog", "cat"]):
+            self.assertEqual(item.annotations[0].label, label_indices[expected])
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_transform_pseudolabeling_without_labels(self):
+        dataset = self.source
+        explorer = Explorer(dataset)
+        result = dataset.transform("pseudo_labeling", explorer=explorer)
+
+        label_indices = dataset.categories()[AnnotationType.label]._indices
+        for item, expected in zip(result, ["dog", "cat"]):
+            self.assertEqual(item.annotations[0].label, label_indices[expected])
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_transform_pseudolabeling_without_explorer(self):
+        dataset = self.source
+        labels = self.categories
+        result = dataset.transform("pseudo_labeling", labels=labels)
+
+        label_indices = dataset.categories()[AnnotationType.label]._indices
+        for item, expected in zip(result, ["dog", "cat"]):
+            self.assertEqual(item.annotations[0].label, label_indices[expected])